+++ code/spikes/publish.001/src/ui/Overlay/useMarkdownLinkClick.mts
@@ -1,5 +1,5 @@
 import { useEffect, useRef } from 'react';
-import { getLinks } from './util.mjs';
+import { visit } from 'unist-util-visit';
 
 import { State, t } from '../common';
 
@@ -50,4 +50,22 @@ export function useMarkdownLinkClick(args: {
    * API
    */
   return { ref };
+}
+
+/**
+ * Helpers
+ */
+
+export function getLinks(mdast?: t.MdastRoot): t.StateOverlayContext[] {
+  if (!mdast) return [];
+  const res: t.StateOverlayContext[] = [];
+
+  visit(mdast, 'link', (node) => {
+    const path = node.url.replace(/^\.\//, '');
+    let title = 'Untitled';
+    if (node.children[0].type === 'text') title = node.children[0].value;
+    res.push({ title, path });
+  });
+
+  return res;
 }
